+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add ability to upgrade firmware during the first-time setup.
 - Change Pacman to use 'curl' instead of 'wget' for downloads.
 - Remove no longer needed workaround to prune NVIDIA runtimes after a Flatpak update.
+- Change the default sleep level to be S3 deep sleep.
 
 ## 4.0.0 - The Conversion Update